+++ .changeset/small-moose-fetch.md
@@ -1,0 +1,11 @@
+---
+'@keystone-next/fields': minor
+'@keystone-next/cloudinary': patch
+'@keystone-next/fields-document': patch
+'@keystone-next/keystone-legacy': patch
+'@keystone-next/server-side-graphql-client-legacy': patch
+'@keystone-next/api-tests-legacy': patch
+'@keystone-next/benchmarks-legacy': patch
+---
+
+Moved the `Implementation` base class from the `fields-legacy` package into the `fields` package.
