+++ packages-next/fields-document/package.json
@@ -26,7 +26,6 @@
     "@keystone-next/admin-ui": "^13.0.0",
     "@keystone-next/admin-ui-utils": "^3.0.3",
     "@keystone-next/fields": "^6.0.0",
-    "@keystone-next/fields-legacy": "^25.0.0",
     "@keystone-next/types": "^16.0.0",
     "@keystone-ui/button": "^3.0.1",
     "@keystone-ui/core": "^2.0.2",
