+++ packages-next/fields/src/Implementation.d.ts
@@ -1,0 +1,4 @@
+// preconstruct(kinda rightfully) gets annoying if you have an import from a ts file
+// to a non-ts file in the same package with no declarations so we need this
+
+export const Implementation: any;
