+++ packages-next/fields/src/types/integer/Implementation.js
@@ -1,5 +1,5 @@
 import { PrismaFieldAdapter } from '@keystone-next/adapter-prisma-legacy';
-import { Implementation } from '@keystone-next/fields-legacy';
+import { Implementation } from '../../Implementation';
 
 export class Integer extends Implementation {
   constructor() {
