+++ packages-next/fields/src/types/relationship/Implementation.js
@@ -1,5 +1,5 @@
 import { PrismaFieldAdapter } from '@keystone-next/adapter-prisma-legacy';
-import { Implementation } from '@keystone-next/fields-legacy';
+import { Implementation } from '../../Implementation';
 import { resolveNested } from './nested-mutations';
 
 export class Relationship extends Implementation {
