+++ packages-next/fields/src/types/select/Implementation.js
@@ -1,6 +1,6 @@
 import inflection from 'inflection';
 import { PrismaFieldAdapter } from '@keystone-next/adapter-prisma-legacy';
-import { Implementation } from '@keystone-next/fields-legacy';
+import { Implementation } from '../../Implementation';
 
 function initOptions(options) {
   let optionsArray = options;
