+++ packages-next/fields/src/types/timestamp/Implementation.js
@@ -1,5 +1,5 @@
 import { PrismaFieldAdapter } from '@keystone-next/adapter-prisma-legacy';
-import { Implementation } from '@keystone-next/fields-legacy';
+import { Implementation } from '../../Implementation';
 
 export class DateTimeUtcImplementation extends Implementation {
   constructor(path, { format = 'yyyy-MM-dd[T]HH:mm:ss.SSSxx' }) {
