--- packages/fields/.npmignore
@@ -1,3 +1,0 @@
-**/*.md
-**/*.test.js
-**/filterTests.js
