--- packages/fields/src/index.js
@@ -1,1 +1,0 @@
-export { Implementation } from './Implementation';
