+++ tests/benchmarks/package.json
@@ -15,7 +15,6 @@
   "homepage": "https://github.com/keystonejs/keystone",
   "dependencies": {
     "@keystone-next/fields": "^6.0.0",
-    "@keystone-next/fields-legacy": "^25.0.0",
     "@keystone-next/keystone": "^15.0.0",
     "@keystone-next/keystone-legacy": "^23.0.0",
     "@keystone-next/server-side-graphql-client-legacy": "^3.0.1",
